Fix events
